pyproject.toml
[project]
name = "cs2smokestelegram"
version = "0.1.0"
description = "A project about the bot sending CS2 lineups"
authors = [
    {name = "Artyom",email = "nechepurenko.work@gmail.com"}
]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-ninja (>=1.4.3,<2.0.0)",
    "aiogram (>=3.21.0,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"
isort = "^6.0.1"
pre-commit = "^4.2.0"
black = "^25.1.0"

[tool.isort]
profile = "black"
skip_glob = ["**/migrations/*.py"]
src_paths = ["src/api", "src/bot",]
